overhauled print processing to use pandas, added pickled file for gui
Altered how prints are stored after detection to use a pandas dataframe, hulls_df, and to store all constituent contours. Changed process_contours and advanced_processing to utilize this dataframe and use pandas to store the final combo_prints. This cleans up code a lot because I can take advantage of built-in pandas functions. I also continue to alter hulls_df as changes are made during advanced_processing. Hulls df and combo prints are saved, hulls df is saved as both a csv (without contours) and a pickled file (with all the contours).